incusd/storage/drivers: refactor ceph discovery
Previously during Ceph storage mounting Incus would try and parse the
`ceph.conf` files itself. While this approach works in the majority of
cases it failed to cover more then a few common setups and was fragile.

This commit changes that to instead makes calls to the Ceph admin tool
`ceph` which provides easy to parse JSON output and is by definition
"correct".

I also introduce a `CephBuildMount` function that essentially takes the
output of `CephMonitors`, `CephFsid`, `CephKeyring`, and the CephFS path
to produce a source path and options list suitable for passing to
`mount`. This also uses the new mount syntax stlye enabling support for
(sub)volumes.

With this the `cephfs.getConfig` function is made redundant and `cephfs`
driver is tweaked to use the modified discovery functions as well as
`CephBuildMount`.

Signed-off-by: Michael 'ASAP' Weinrich <michael@a5ap.net>